Add permanent docker-worker public cot key r=aki
So far, each time we build a new AMI for docker-worker trusted
worker-types, we generate the cot signing key pair during AMI creation,
extract the public key and add it to the cot repo. The original mindset
is that the private key should never be in anyone's hands. A careful
threat modeling shows that the private key is exposed to anyone with AWS
credentials. Conclusion: the private is as safe as the AWS credentials
are. Therefore, generating a new key pair each time a AMI is built is
useless, and adds complexity to AMI rollout (push pub key to cot repo)
for no reward.

That said, docker-worker deploy is moving to use a permanent signing key
that is changed only on special cases (e.g., the key is leaked). The
private key will be stored to in passwordstore, while the pub is added
to cot in this commit.